Close the modal when the form is submitted
The modal used to be closed when add book button
was clicked, regardless if the form was filled or not.
Now, the form has to be filled in order for the button
to close the modal.
If users wish to close the modal without filling
the form, they can use the closing button.